bugfix: prevent a NULL pointer dereference
pAction->ppTpl was dereferenced if
(iTplOpts & OMSR_TPL_AS_MSG) && (iTplOpts & OMSR_RQD_TPL_OPT_SQL)

The pAction->eParamPassing variable sets behavior for the whole action
and all templates, but the option variable used to set it is
per-template - this can lead to issues if there are conflicts. The
code was refactored to make it more readable and to set eParamPassing
only once. pAction->ppTpl is only allocated when needed.